+++ NEWS.md
@@ -1,13 +1,15 @@
 iptools 0.2.0 [WIP]
 =============
 NEW FEATURES
-* gethostbyname and gethostsbynames are now the same function, named hostname\_to\_dns
+* gethostbyname and gethostsbynames are now the same function, named hostname\_to\_ip
+* gethostbyaddr and gethostsbyaddr are now the same function, named ip_to_hostname
 
 BUG FIXES
 
 DEVELOPMENT
 *geolocation functionality deprecated; rely on install_github("ironholds/rgeoip") for geolocation.
 *hostname-to-DNS code, including the vectorisation, moved to C++
+
 iptools 0.1.4
 =============